Python lighting script
Due to how the Python for Unity API is implemented, we cant pass any
additional arguments to the script when calling it fully via Unity. As
such, an additional way to interact with the scrip is added to overcome
this limitation.

Add lighting documentation
Added possible scene names to the README